Upload Files: House Tracker Simulator
The House Tracker Simulator receives an input of a csv file with hundreds of thousands of houses for sale. Specifically, it includes specific dimensions from these houses such as its length, width, pool size, garage space, and more. We use linear regression as well as data extraction techniques to find the best house suitable for the user's needs and preferences.