Upload Files: Nearest Neighbors Projects
These projects revolve around the idea of Nearest Neighbors in Python. 

One project is the Digit Identifier where I designed a ML model that accepts an image of a blurry or masked digit and outputs its prediction for what that digit is in clarity. I developed training and testing models to improve accuracy of prediction results. I obtained the pixel values of the digit and compared those to the pixel values of the clear digit. 

I utilized a wide range of datasets in the form of csv files. I implemented pandas and matplotlib for perceptual ease and to gain more information about the data and results.

I used the same algorithms to find the most common birthdays in a given set of fictional participants. The input is in the form of a csv file that contains the birthdays of hundreds of thousands of participants. 

I have also created a Species Predictor where I designed a model that determines species of various flowers given different dimensions including petal and sepal lengths and widths. The data inputted was in the form of a csv file with hundreds of thousands of different values for various categories of flower components.

Lastly, I branched out and used these algorithms for a given set of Titanic passengers. Our model predicts whether the passenger is deceased from the Titanic crash or not. We give our models information about the passengers' age, sex, number of siblings, room number, and much more.